requirements.txt
apipkg==1.5
asgiref==3.3.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
coverage==5.5
Django==3.1.7
django-cors-headers==3.7.0
django-debug-toolbar==3.2
django-redis==4.12.1
djangorestframework==3.12.4
execnet==1.8.0
gunicorn==20.0.4
idna==2.10
iniconfig==1.1.1
mbstrdecoder==1.0.1
packaging==20.9
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
pytest-cov==2.11.1
pytest-deadfixtures==2.2.1
pytest-django==4.1.0
pytest-dotenv==0.5.2
pytest-forked==1.3.0
pytest-mock==3.5.1
pytest-xdist==2.2.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typepy==1.1.4
urllib3==1.26.4
